+++ server/api/src/main/java/com/strongkey/apiws/filters/CrossOriginResourceSharingFilter.java
@@ -4,7 +4,7 @@
  * Use of this source code is governed by the Gnu Lesser General Public License 2.3.
  * The license can be found at https://github.com/StrongKey/fido2/LICENSE
  */
-package com.strongauth.apiws.filters;
+package com.strongkey.apiws.filters;
 
 import com.strongkey.appliance.utilities.applianceCommon;
 import java.io.IOException;
