+++ server/fidoserverbeans/pom.xml
@@ -3,12 +3,12 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <artifactId>fidoserver</artifactId>
-        <groupId>com.strongauth</groupId>
+        <groupId>com.strongkey</groupId>
         <version>4.0</version>
          <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <groupId>com.strongauth</groupId>
+    <groupId>com.strongkey</groupId>
     <artifactId>fidoserverbeans</artifactId>
     <version>4.0</version>
     <packaging>ejb</packaging>
@@ -114,18 +114,16 @@
             <scope>provided</scope>
         </dependency>
         <dependency>
-            <groupId>com.strongauth</groupId>
+            <groupId>com.strongkey</groupId>
             <artifactId>replicationWAR</artifactId>
-            <version>4.0</version>
-            <type>war</type>
+            <version>4.1.0</version>
             <scope>provided</scope>
         </dependency>
         
         <dependency>
-            <groupId>com.strongauth</groupId>
+            <groupId>com.strongkey</groupId>
             <artifactId>replicationEJB</artifactId>
-            <version>4.0</version>
-            <type>ejb</type>
+            <version>4.1.0</version>
             <scope>provided</scope>
         </dependency>
         
@@ -141,11 +139,11 @@
             <version>4.0</version>
             <exclusions>
                 <exclusion>
-                    <groupId>com.strongauth</groupId>
+                    <groupId>com.strongkey</groupId>
                     <artifactId>common</artifactId>
                 </exclusion>
                 <exclusion>
-                    <groupId>com.strongauth</groupId>
+                    <groupId>com.strongkey</groupId>
                     <artifactId>crypto</artifactId>
                 </exclusion>
             </exclusions>
