+++ server/skfe/src/main/java/com/strongkey/skfews/u2f/rest/ApplicationConfig.java
@@ -1,4 +1,4 @@
-/*
+/**
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
  * License, as published by the Free Software Foundation and
@@ -12,10 +12,10 @@
  *
  * Copyright (c) 2001-2018 StrongAuth, Inc.
  *
- * $Date: 
- * $Revision:
- * $Author: mishimoto $
- * $URL: 
+ * $Date$
+ * $Revision$
+ * $Author$
+ * $URL$
  *
  * *********************************************
  *                    888
@@ -28,30 +28,23 @@
  *  888  888  "Y88P"   "Y888  "Y8888   88888P'
  *
  * *********************************************
- * 
  *
+ * ApplicationConfig class to define RESTful interface for SKFEServlet class.
  *
  */
-package com.strongauth.skfews.fido2.rest;
+package com.strongkey.skfews.u2f.rest;
 
-import javax.json.Json;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+import javax.ws.rs.ApplicationPath;
+import javax.ws.rs.core.Application;
 
-public class FIDOResponse {
-    private final String response;
-    private final String message;
-    private final String error;
-    
-    public FIDOResponse(String response, String message, String error){
-        this.response = response;
-        this.message = message;
-        this.error = error;
-    }
-    
+@ApplicationPath("/rest")
+public class ApplicationConfig extends Application {
+
     @Override
-    public String toString(){
-        return Json.createObjectBuilder()
-                .add("Response", response)
-                .add("Message", message)
-                .add("Error", error).build().toString();
+    public Set<Class<?>> getClasses() {
+        return new HashSet<>(Arrays.asList(SKFEServlet.class));
     }
 }
