+++ server/skfe/src/main/java/com/strongkey/skfews/u2f/soap/SKFEServlet.java
@@ -1,20 +1,21 @@
 /*
- * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License, as published by the Free
- * Software Foundation and available at
- * http://www.fsf.org/licensing/licenses/lgpl.html, version 2.1 or above.
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License, as published by the Free Software Foundation and
+ * available at http://www.fsf.org/licensing/licenses/lgpl.html,
+ * version 2.1 or above.
  *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
  *
  * Copyright (c) 2001-2018 StrongAuth, Inc.
  *
- * $Date$ $Revision$
- * $Author$ $URL:
- * https://svn.strongauth.com/repos/jade/trunk/skce/skfe/src/main/java/com/strongauth/skfews/u2f/rest/SKFEServlet.java
- * $
+ * $Date$
+ * $Revision$
+ * $Author$
+ * $URL$
  *
  * *********************************************
  *                    888
@@ -28,128 +29,66 @@
  *
  * *********************************************
  *
- * Servlet for FIDO U2F protocol based functionality. This servlet exposes REST
- * (Representational State Transfer) based web services to the calling
+ * Servlet for FIDO U2F protocol based functionality. This servlet exposes 
+ * SOAP (Simple Object Access Protocol) based web services to the calling 
  * applications.
  *
  */
-package com.strongauth.skfews.u2f.rest;
+package com.strongkey.skfews.u2f.soap;
 
-
-import com.strongkey.appliance.utilities.applianceCommon;
 import com.strongkey.auth.txbeans.authorizeLdapUserBeanLocal;
 import com.strongkey.skce.jaxb.SKCEServiceInfoType;
+import com.strongkey.skce.utilities.SKCEException;
 import com.strongkey.skfs.txbeans.v1.u2fServletHelperBeanLocal_v1;
-import com.strongkey.skfs.utilities.SKFEException;
 import com.strongkey.skfs.utilities.skfsCommon;
 import com.strongkey.skfs.utilities.skfsConstants;
 import com.strongkey.skfs.utilities.skfsLogger;
 import java.util.logging.Level;
 import javax.ejb.EJB;
-import javax.naming.Context;
-import javax.naming.InitialContext;
-import javax.naming.NamingException;
+import javax.jws.WebMethod;
+import javax.jws.WebParam;
+import javax.jws.WebService;
 import javax.servlet.http.HttpServletRequest;
-import javax.ws.rs.Consumes;
-import javax.ws.rs.FormParam;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
 
 /**
- * REST based web services that serve FIDO U2F protocol based functionality.
+ * SOAP based web services that serve FIDO U2F protocol based functionality.
  *
  */
-@Path("")
+@WebService(serviceName = "soap")
 public class SKFEServlet {
 
-    @EJB
-    authorizeLdapUserBeanLocal authorizebean = lookupauthorizeLdapUserBeanLocal();
-
-    String skfehome = skfsCommon.getSkfeHome();
-
-    @javax.ws.rs.core.Context
+    @Context
     private HttpServletRequest request;
-
-    /*
-     * Enterprise Java Beans used in this servlet.
-     */
+    
     @EJB
-    u2fServletHelperBeanLocal_v1 u2fHelperBean = lookup_u2fServletHelperBeanLocal_v1();
-
-    /**
-     * methods to look up for ejb resources
-     */
-    private u2fServletHelperBeanLocal_v1 lookup_u2fServletHelperBeanLocal_v1() {
-        try {
-            Context c = new InitialContext();
-            return (u2fServletHelperBeanLocal_v1) c.lookup("java:app/fidoserverbeans-4.0/u2fServletHelperBean_v1!com.strongkey.skfs.txbeans.v1.u2fServletHelperBeanLocal_v1");
-        } catch (NamingException ne) {
-            throw new RuntimeException(ne);
-        }
-    }
-
-    private authorizeLdapUserBeanLocal lookupauthorizeLdapUserBeanLocal() {
-        try {
-            Context c = new InitialContext();
-            return (authorizeLdapUserBeanLocal) c.lookup("java:app/authenticationBeans-4.0/authorizeLdapUserBean!com.strongkey.auth.txbeans.authorizeLdapUserBeanLocal");
-        } catch (NamingException ne) {
-            throw new RuntimeException(ne);
-        }
-    }
-
+    u2fServletHelperBeanLocal_v1 u2fHelper;
+    @EJB
+    authorizeLdapUserBeanLocal authorizebean; // ldap user authorization bean 
     // CTOR
+
     public SKFEServlet() {
     }
     
     /**
      * Basic null checks before the compound input objects are accessed.
      * 
-     * @param svcinfo String; Json string with service credentials information
+     * @param svcinfo       SKCEServiceInfoType; service credentials information
      * 
-     * @return SKCEServiceInfoType; if all checks passed
-     * @throws SKFEException; in case any check fails
+     * @return void; if all checks passed
+     * @throws SKCEException; in case any check fails
      */
-    private SKCEServiceInfoType basicInputChecks(String methodname,
-            String svcinfo) throws SKFEException {
+    private void basicInputChecks(String methodname, 
+                                  SKCEServiceInfoType svcinfo) throws SKCEException 
+    {       
         String prefix = methodname + " web-service; ";
-
-        if (svcinfo == null || svcinfo.trim().isEmpty()) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER, Level.SEVERE, "SKCEWS-ERR-3014",
+        
+        if ( svcinfo==null ) {
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "SKCEWS-ERR-3014", 
                     "svcinfo");
-            throw new SKFEException(skfsCommon.getMessageProperty("SKCEWS-ERR-3014")
-                    .replace("{0}", "") + prefix + "svcinfo");
-        } else if (!skfsCommon.isValidJsonObject(svcinfo)) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER, Level.SEVERE, "FIDO-ERR-0014",
-                    prefix + "Invalid json; svcinfo");
-            throw new SKFEException(skfsCommon.getMessageProperty("FIDO-ERR-0014")
-                    .replace("{0}", "") + prefix + "Invalid json; svcinfo");
-        }
-
-        try {
-            //  Parse out the needed key-values from json
-            String did = (String) applianceCommon.getJsonValue(svcinfo, "did", "String");
-            String svcusername = (String) applianceCommon.getJsonValue(svcinfo, "svcusername", "String");
-            String svcpassword = (String) applianceCommon.getJsonValue(svcinfo, "svcpassword", "String");
-            String protocol = (String) applianceCommon.getJsonValue(svcinfo, "protocol", "String");
-
-            //  Construct serviceinfo object
-            SKCEServiceInfoType si = new SKCEServiceInfoType();
-            si.setDid(Integer.parseInt(did));
-            si.setSvcusername(svcusername);
-            si.setSvcpassword(svcpassword);
-            if (protocol != null && !protocol.trim().isEmpty()) {
-                si.setProtocol(protocol);
-            }
-            return si;
-
-        } catch (Exception ex) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER, Level.SEVERE, "SKCEWS-ERR-3053",
-                    "svcinfo");
-            throw new SKFEException(skfsCommon.getMessageProperty("SKCEWS-ERR-3053")
+            throw new SKCEException(skfsCommon.getMessageProperty("SKCEWS-ERR-3014")
                     .replace("{0}", "") + prefix + "svcinfo");
         }
-
     }
 
     /*
@@ -172,29 +111,32 @@ public class SKFEServlet {
      * challenge and returns the same to the caller, which typically is a
      * Relying Party (RP) application.
      *
-     * @param svcinfo - Object that carries SKCE service information.
-     * Information bundled is :
+     * @param svcinfo   - Object that carries SKCE service information. 
+     *                    Information bundled is :
      * 
-     * (1) did - Unique identifier for a SKCE encryption domain (2) svcusername
-     * - SKCE service credentials : username requesting the service. The service
-     * credentials are looked up in the 'service' setup of authentication system
-     * based on LDAP / AD. The user should be authorized to encrypt. (3)
-     * svcpassword - SKCE service credentials : password of the service username
-     * specified above (4) protocol - U2F protocol version to comply with.
+     * (1) did          - Unique identifier for a SKCE encryption domain
+     * (2) svcusername  - SKCE service credentials : username requesting the 
+     *                    service. The service credentials are looked up in 
+     *                    the 'service' setup of authentication system based 
+     *                    on LDAP / AD.
+     *                    The user should be authorized to encrypt.
+     * (3) svcpassword -  SKCE service credentials : password of the service 
+     *                    username specified above
+     * (4) protocol    -  U2F protocol version to comply with.
      * 
-     * @param payload
+     * @param payload - payload
+     * 
      * @return - A Json in String format. The Json will have 3 key-value pairs;
-     * 1. 'Challenge' : 'U2F Reg Challenge parameters; a json again' 2.
-     * 'Message' : String, with a list of messages that explain the process. 3.
-     * 'Error' : String, with error message incase something went wrong. Will be
-     * empty if successful.
+     * 1. 'Challenge' : U2F Reg Challenge parameters; a json again 2. 'Message'
+     * : String, with a list of messages that explain the process. 3. 'Error' :
+     * String, with error message incase something went wrong. Will be empty if
+     * successful.
      */
-    @POST
-    @Path("/" + skfsConstants.FIDO_METHOD_PREREGISTER)
-    @Consumes({"application/x-www-form-urlencoded"})
-    @Produces({"application/json"})
-    public String preregister(@FormParam("svcinfo") String svcinfo,
-            @FormParam("payload") String payload) {
+    @WebMethod(operationName = "preregister")
+    public String preregister(
+            @WebParam(name = "svcinfo")  SKCEServiceInfoType svcinfo,
+            @WebParam(name = "payload") String payload) 
+    {
         //  Local variables       
         //  Service credentials
         String did;
@@ -204,29 +146,28 @@ public class SKFEServlet {
         
         //  SKCE domain id validation
         try {
-            SKCEServiceInfoType si = basicInputChecks("preregister", svcinfo);
+            basicInputChecks("preregister", svcinfo);
             
-            did = Integer.toString(si.getDid());
-            svcusername = si.getSvcusername();
-            svcpassword = si.getSvcpassword();
-            protocol = si.getProtocol();
+            did             = Integer.toString(svcinfo.getDid());
+            svcusername     = svcinfo.getSvcusername();
+            svcpassword     = svcinfo.getSvcpassword();
+            protocol        = svcinfo.getProtocol();
 
             skfsCommon.inputValidateSKCEDid(did);
         } catch (Exception ex) {
             return skfsCommon.buildPreRegisterResponse(null, "", ex.getLocalizedMessage());
         }
         
-        //  Service credentials input checks
+        //  2. Input checks
         if (svcusername == null || svcusername.isEmpty()) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcusername");
-            return skfsCommon.buildPreRegisterResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcusername");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildPreRegisterResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         if (svcpassword == null) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcpassword");
-            return skfsCommon.buildPreRegisterResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcpassword");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildPreRegisterResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
-        
-        // Service credentials' authentication and authorization
+        //authenticate
         boolean isAuthorized;
         try {
             isAuthorized = authorizebean.execute(Long.parseLong(did), svcusername, svcpassword, skfsConstants.LDAP_ROLE_FIDO);
@@ -238,7 +179,8 @@ public class SKFEServlet {
             skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0033", "");
             return skfsCommon.buildPreRegisterResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0033"));
         }
-        return u2fHelperBean.preregister(did, protocol, payload);
+        
+        return u2fHelper.preregister(did, protocol, payload);
     }
 
     /*
@@ -265,30 +207,32 @@ public class SKFEServlet {
      * should happen with in a certain time limit after the preregister is
      * finished; otherwise, the user session would be invalidated.
      *
-     * @param svcinfo - Object that carries SKCE service information.
-     * Information bundled is :
+     * @param svcinfo   - Object that carries SKCE service information. 
+     *                    Information bundled is :
      * 
-     * (1) did - Unique identifier for a SKCE encryption domain (2) svcusername
-     * - SKCE service credentials : username requesting the service. The service
-     * credentials are looked up in the 'service' setup of authentication system
-     * based on LDAP / AD. The user should be authorized to encrypt. (3)
-     * svcpassword - SKCE service credentials : password of the service username
-     * specified above
+     * (1) did          - Unique identifier for a SKCE encryption domain
+     * (2) svcusername  - SKCE service credentials : username requesting the 
+     *                    service. The service credentials are looked up in 
+     *                    the 'service' setup of authentication system based 
+     *                    on LDAP / AD.
+     *                    The user should be authorized to encrypt.
+     * (3) svcpassword -  SKCE service credentials : password of the service 
+     *                    username specified above
      * 
      * @param payload - U2F Registration Response parameters in Json form.
      * Should contain sessionid, browserData and enrollData.
+     * 
      * @return - A Json in String format. The Json will have 3 key-value pairs;
      * 1. 'Response' : String, with a simple message telling if the process was
      * successful or not. 2. 'Message' : String, with a list of messages that
      * explain the process. 3. 'Error' : String, with error message incase
      * something went wrong. Will be empty if successful.
      */
-    @POST
-    @Path("/" + skfsConstants.FIDO_METHOD_REGISTER)
-    @Consumes({"application/x-www-form-urlencoded"})
-    @Produces({"application/json"})
-    public String register(@FormParam("svcinfo") String svcinfo,
-            @FormParam("payload") String payload) {
+    @WebMethod(operationName = "register")
+    public String register(
+            @WebParam(name = "svcinfo")  SKCEServiceInfoType svcinfo,
+            @WebParam(name = "payload") String payload) 
+    {
         //  Local variables       
         //  Service credentials
         String did;
@@ -298,24 +242,26 @@ public class SKFEServlet {
         
         //  SKCE domain id validation
         try {
-            SKCEServiceInfoType si = basicInputChecks("register", svcinfo);
-            did = Integer.toString(si.getDid());
-            svcusername = si.getSvcusername();
-            svcpassword = si.getSvcpassword();
-            protocol = si.getProtocol();
+            basicInputChecks("register", svcinfo);
+            
+            did             = Integer.toString(svcinfo.getDid());
+            svcusername     = svcinfo.getSvcusername();
+            svcpassword     = svcinfo.getSvcpassword();
+            protocol        = svcinfo.getProtocol();
             
             skfsCommon.inputValidateSKCEDid(did);
         } catch (Exception ex) {
             return skfsCommon.buildRegisterResponse(null, "", ex.getLocalizedMessage());
         }
         
+        //  2. Input checks
         if (svcusername == null || svcusername.isEmpty()) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcusername");
-            return skfsCommon.buildRegisterResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcusername");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildRegisterResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         if (svcpassword == null) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcpassword");
-            return skfsCommon.buildRegisterResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcpassword");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildRegisterResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         //authenticate
         boolean isAuthorized;
@@ -329,8 +275,7 @@ public class SKFEServlet {
             skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0033", "");
             return skfsCommon.buildRegisterResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0033"));
         }
-        
-        return u2fHelperBean.register(did, protocol, payload);
+        return u2fHelper.register(did, protocol, payload);
     }
 
     /*
@@ -352,29 +297,32 @@ public class SKFEServlet {
      * Step-1 for fido authenticator authentication. This methods generates a
      * challenge and returns the same to the caller.
      *
-     * @param svcinfo - Object that carries SKCE service information.
-     * Information bundled is :
+     * @param svcinfo   - Object that carries SKCE service information. 
+     *                    Information bundled is :
      * 
-     * (1) did - Unique identifier for a SKCE encryption domain (2) svcusername
-     * - SKCE service credentials : username requesting the service. The service
-     * credentials are looked up in the 'service' setup of authentication system
-     * based on LDAP / AD. The user should be authorized to encrypt. (3)
-     * svcpassword - SKCE service credentials : password of the service username
-     * specified above (4) protocol - U2F protocol version to comply with.
+     * (1) did          - Unique identifier for a SKCE encryption domain
+     * (2) svcusername  - SKCE service credentials : username requesting the 
+     *                    service. The service credentials are looked up in 
+     *                    the 'service' setup of authentication system based 
+     *                    on LDAP / AD.
+     *                    The user should be authorized to encrypt.
+     * (3) svcpassword -  SKCE service credentials : password of the service 
+     *                    username specified above
+     * (4) protocol    -  U2F protocol version to comply with.
      * 
      * @param payload - username
+     * 
      * @return - A Json in String format. The Json will have 3 key-value pairs;
      * 1. 'Challenge' : 'U2F Auth Challenge parameters; a json again' 2.
      * 'Message' : String, with a list of messages that explain the process. 3.
      * 'Error' : String, with error message incase something went wrong. Will be
      * empty if successful.
      */
-    @POST
-    @Path("/" + skfsConstants.FIDO_METHOD_PREAUTH)
-    @Consumes({"application/x-www-form-urlencoded"})
-    @Produces({"application/json"})
-    public String preauthenticate(@FormParam("svcinfo") String svcinfo,
-            @FormParam("payload") String payload) {
+    @WebMethod(operationName = "preauthenticate")
+    public String preauthenticate(
+            @WebParam(name = "svcinfo")  SKCEServiceInfoType svcinfo,
+            @WebParam(name = "payload") String payload) 
+    {
         //  Local variables       
         //  Service credentials
         String did;
@@ -384,12 +332,13 @@ public class SKFEServlet {
         
         //  SKCE domain id validation
         try {
-            SKCEServiceInfoType si = basicInputChecks("preauthenticate", svcinfo);
-            did = Integer.toString(si.getDid());
-            svcusername = si.getSvcusername();
-            svcpassword = si.getSvcpassword();
-            protocol = si.getProtocol();
-
+            basicInputChecks("preauthenticate", svcinfo);
+            
+            did             = Integer.toString(svcinfo.getDid());
+            svcusername     = svcinfo.getSvcusername();
+            svcpassword     = svcinfo.getSvcpassword();
+            protocol        = svcinfo.getProtocol();
+            
             skfsCommon.inputValidateSKCEDid(did);
         } catch (Exception ex) {
             return skfsCommon.buildPreAuthResponse(null, "", ex.getLocalizedMessage());
@@ -397,12 +346,12 @@ public class SKFEServlet {
         
         //  2. Input checks
         if (svcusername == null || svcusername.isEmpty()) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcusername");
-            return skfsCommon.buildPreAuthResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcusername");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildPreAuthResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         if (svcpassword == null) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcpassword");
-            return skfsCommon.buildPreAuthResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcpassword");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildPreAuthResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         //authenticate
         boolean isAuthorized;
@@ -416,7 +365,7 @@ public class SKFEServlet {
             skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0033", "");
             return skfsCommon.buildPreAuthResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0033"));
         }
-        return u2fHelperBean.preauthenticate(did, protocol, payload);
+        return u2fHelper.preauthenticate(did, protocol, payload);
     }
 
     /*
@@ -442,15 +391,17 @@ public class SKFEServlet {
      * preauthenticate is finished; otherwise, the user session would be
      * invalidated.
      *
-     * @param svcinfo - Object that carries SKCE service information.
-     * Information bundled is :
+     * @param svcinfo   - Object that carries SKCE service information. 
+     *                    Information bundled is :
      * 
-     * (1) did - Unique identifier for a SKCE encryption domain (2) svcusername
-     * - SKCE service credentials : username requesting the service. The service
-     * credentials are looked up in the 'service' setup of authentication system
-     * based on LDAP / AD. The user should be authorized to encrypt. (3)
-     * svcpassword - SKCE service credentials : password of the service username
-     * specified above
+     * (1) did          - Unique identifier for a SKCE encryption domain
+     * (2) svcusername  - SKCE service credentials : username requesting the 
+     *                    service. The service credentials are looked up in 
+     *                    the 'service' setup of authentication system based 
+     *                    on LDAP / AD.
+     *                    The user should be authorized to encrypt.
+     * (3) svcpassword -  SKCE service credentials : password of the service 
+     *                    username specified above
      * 
      * @param payload
      * @return - A Json in String format. The Json will have 3 key-value pairs;
@@ -459,12 +410,11 @@ public class SKFEServlet {
      * explain the process. 3. 'Error' : String, with error message incase
      * something went wrong. Will be empty if successful.
      */
-    @POST
-    @Path("/" + skfsConstants.FIDO_METHOD_AUTHENTICATE)
-    @Consumes({"application/x-www-form-urlencoded"})
-    @Produces({"application/json"})
-    public String authenticate(@FormParam("svcinfo") String svcinfo,
-            @FormParam("payload") String payload) {
+    @WebMethod(operationName = "authenticate")
+    public String authenticate(
+            @WebParam(name = "svcinfo")  SKCEServiceInfoType svcinfo,
+            @WebParam(name = "payload") String payload) 
+    {
         //  Local variables       
         //  Service credentials
         String did;
@@ -474,12 +424,13 @@ public class SKFEServlet {
         
         //  SKCE domain id validation
         try {
-            SKCEServiceInfoType si = basicInputChecks("authenticate", svcinfo);
-            did = Integer.toString(si.getDid());
-            svcusername = si.getSvcusername();
-            svcpassword = si.getSvcpassword();
-            protocol = si.getProtocol();
-
+            basicInputChecks("authenticate", svcinfo);
+            
+            did             = Integer.toString(svcinfo.getDid());
+            svcusername     = svcinfo.getSvcusername();
+            svcpassword     = svcinfo.getSvcpassword();
+            protocol        = svcinfo.getProtocol();
+            
             skfsCommon.inputValidateSKCEDid(did);
         } catch (Exception ex) {
             return skfsCommon.buildAuthenticateResponse(null, "", ex.getLocalizedMessage());
@@ -487,12 +438,12 @@ public class SKFEServlet {
         
         //  2. Input checks
         if (svcusername == null || svcusername.isEmpty()) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcusername");
-            return skfsCommon.buildAuthenticateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcusername");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildAuthenticateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         if (svcpassword == null) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcpassword");
-            return skfsCommon.buildAuthenticateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcpassword");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildAuthenticateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         //authenticate
         boolean isAuthorized;
@@ -506,7 +457,7 @@ public class SKFEServlet {
             skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0033", "");
             return skfsCommon.buildAuthenticateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0033"));
         }
-        return u2fHelperBean.authenticate(did, protocol, payload);
+        return u2fHelper.authenticate(did, protocol, payload);
     }
 
     /*
@@ -528,15 +479,18 @@ public class SKFEServlet {
      * Step-1 for fido based transaction confirmation using u2f authenticator.
      * This methods generates a challenge and returns the same to the caller.
      *
-     * @param svcinfo - Object that carries SKCE service information.
-     * Information bundled is :
+     * @param svcinfo   - Object that carries SKCE service information. 
+     *                    Information bundled is :
      * 
-     * (1) did - Unique identifier for a SKCE encryption domain (2) svcusername
-     * - SKCE service credentials : username requesting the service. The service
-     * credentials are looked up in the 'service' setup of authentication system
-     * based on LDAP / AD. The user should be authorized to encrypt. (3)
-     * svcpassword - SKCE service credentials : password of the service username
-     * specified above (4) protocol - U2F protocol version to comply with.
+     * (1) did          - Unique identifier for a SKCE encryption domain
+     * (2) svcusername  - SKCE service credentials : username requesting the 
+     *                    service. The service credentials are looked up in 
+     *                    the 'service' setup of authentication system based 
+     *                    on LDAP / AD.
+     *                    The user should be authorized to encrypt.
+     * (3) svcpassword -  SKCE service credentials : password of the service 
+     *                    username specified above
+     * (4) protocol    -  U2F protocol version to comply with.
      * 
      * @param payload - String indicating transaction reference.
      * @return - A Json in String format. The Json will have 3 key-value pairs;
@@ -545,12 +499,11 @@ public class SKFEServlet {
      * 'Error' : String, with error message incase something went wrong. Will be
      * empty if successful.
      */
-    @POST
-    @Path("/" + skfsConstants.FIDO_METHOD_PREAUTHORIZE)
-    @Consumes({"application/x-www-form-urlencoded"})
-    @Produces({"application/json"})
-    public String preauthorize(@FormParam("svcinfo") String svcinfo,
-            @FormParam("payload") String payload) {
+    @WebMethod(operationName = "preauthorize")
+    public String preauthorize(
+            @WebParam(name = "svcinfo")  SKCEServiceInfoType svcinfo,
+            @WebParam(name = "payload") String payload) 
+    {
         //  Local variables       
         //  Service credentials
         String did;
@@ -560,12 +513,13 @@ public class SKFEServlet {
         
         //  SKCE domain id validation
         try {
-            SKCEServiceInfoType si = basicInputChecks("preauthorize", svcinfo);
-            did = Integer.toString(si.getDid());
-            svcusername = si.getSvcusername();
-            svcpassword = si.getSvcpassword();
-            protocol = si.getProtocol();
-
+            basicInputChecks("preauthorize", svcinfo);
+            
+            did             = Integer.toString(svcinfo.getDid());
+            svcusername     = svcinfo.getSvcusername();
+            svcpassword     = svcinfo.getSvcpassword();
+            protocol        = svcinfo.getProtocol();
+            
             skfsCommon.inputValidateSKCEDid(did);
         } catch (Exception ex) {
             return skfsCommon.buildPreAuthResponse(null, "", ex.getLocalizedMessage());
@@ -573,12 +527,12 @@ public class SKFEServlet {
         
         //  2. Input checks
         if (svcusername == null || svcusername.isEmpty()) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcusername");
-            return skfsCommon.buildPreAuthResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcusername");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildPreAuthResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         if (svcpassword == null) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcpassword");
-            return skfsCommon.buildPreAuthResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcpassword");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildPreAuthResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         //authenticate
         boolean isAuthorized;
@@ -592,7 +546,7 @@ public class SKFEServlet {
             skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0033", "");
             return skfsCommon.buildPreAuthResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0033"));
         }
-        return u2fHelperBean.preauthorize(did, protocol, payload);
+        return u2fHelper.preauthorize(did, protocol, payload);
     }
 
     /*
@@ -619,29 +573,30 @@ public class SKFEServlet {
      * preauthorize is finished; otherwise, the user session would be
      * invalidated.
      *
-     * @param svcinfo - Object that carries SKCE service information.
-     * Information bundled is :
+     * @param svcinfo   - Object that carries SKCE service information. 
+     *                    Information bundled is :
      * 
-     * (1) did - Unique identifier for a SKCE encryption domain (2) svcusername
-     * - SKCE service credentials : username requesting the service. The service
-     * credentials are looked up in the 'service' setup of authentication system
-     * based on LDAP / AD. The user should be authorized to encrypt. (3)
-     * svcpassword - SKCE service credentials : password of the service username
-     * specified above
-     *
-     * @param payload * @return - A Json in String format. The Json will have 3
-     * key-value pairs; 1. 'Response' : String, with a simple message telling if
-     * the process was successful or not. 2. 'Message' : String, with a list of
-     * messages that explain the process. 3. 'Error' : String, with error
-     * message incase something went wrong. Will be empty if successful.
-     * @return 
+     * (1) did          - Unique identifier for a SKCE encryption domain
+     * (2) svcusername  - SKCE service credentials : username requesting the 
+     *                    service. The service credentials are looked up in 
+     *                    the 'service' setup of authentication system based 
+     *                    on LDAP / AD.
+     *                    The user should be authorized to encrypt.
+     * (3) svcpassword -  SKCE service credentials : password of the service 
+     *                    username specified above
+     * 
+     * @param payload
+     * @return - A Json in String format. The Json will have 3 key-value pairs;
+     * 1. 'Response' : String, with a simple message telling if the process was
+     * successful or not. 2. 'Message' : String, with a list of messages that
+     * explain the process. 3. 'Error' : String, with error message incase
+     * something went wrong. Will be empty if successful.
      */
-    @POST
-    @Path("/" + skfsConstants.FIDO_METHOD_AUTHORIZE)
-    @Consumes({"application/x-www-form-urlencoded"})
-    @Produces({"application/json"})
-    public String authorize(@FormParam("svcinfo") String svcinfo,
-            @FormParam("payload") String payload) {
+    @WebMethod(operationName = "authorize")
+    public String authorize(
+            @WebParam(name = "svcinfo") SKCEServiceInfoType svcinfo,
+            @WebParam(name = "payload") String payload) 
+    {
         //  Local variables       
         //  Service credentials
         String did;
@@ -651,13 +606,13 @@ public class SKFEServlet {
         
         //  SKCE domain id validation
         try {
-            SKCEServiceInfoType si = basicInputChecks("authorize", svcinfo);
+            basicInputChecks("authorize", svcinfo);
             
-            did = Integer.toString(si.getDid());
-            svcusername = si.getSvcusername();
-            svcpassword = si.getSvcpassword();
-            protocol = si.getProtocol();
-
+            did             = Integer.toString(svcinfo.getDid());
+            svcusername     = svcinfo.getSvcusername();
+            svcpassword     = svcinfo.getSvcpassword();
+            protocol        = svcinfo.getProtocol();
+            
             skfsCommon.inputValidateSKCEDid(did);
         } catch (Exception ex) {
             return skfsCommon.buildAuthenticateResponse(null, "", ex.getLocalizedMessage());
@@ -665,12 +620,12 @@ public class SKFEServlet {
         
         //  2. Input checks
         if (svcusername == null || svcusername.isEmpty()) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcusername");
-            return skfsCommon.buildAuthenticateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcusername");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildAuthenticateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         if (svcpassword == null) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcpassword");
-            return skfsCommon.buildAuthenticateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcpassword");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildAuthenticateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         //authenticate
         boolean isAuthorized;
@@ -684,7 +639,7 @@ public class SKFEServlet {
             skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0033", "");
             return skfsCommon.buildAuthenticateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0033"));
         }
-        return u2fHelperBean.authorize(did, protocol, payload);
+        return u2fHelper.authorize(did, protocol, payload);
     }
 
     /*
@@ -709,30 +664,33 @@ public class SKFEServlet {
      * id to point to a unique registered key for that user. This random id can
      * be obtained by calling getkeysinfo method.
      *
-     * @param svcinfo - Object that carries SKCE service information.
-     * Information bundled is :
+     * @param svcinfo   - Object that carries SKCE service information. 
+     *                    Information bundled is :
      * 
-     * (1) did - Unique identifier for a SKCE encryption domain (2) svcusername
-     * - SKCE service credentials : username requesting the service. The service
-     * credentials are looked up in the 'service' setup of authentication system
-     * based on LDAP / AD. The user should be authorized to encrypt. (3)
-     * svcpassword - SKCE service credentials : password of the service username
-     * specified above (4) protocol - U2F protocol version to comply with.
+     * (1) did          - Unique identifier for a SKCE encryption domain
+     * (2) svcusername  - SKCE service credentials : username requesting the 
+     *                    service. The service credentials are looked up in 
+     *                    the 'service' setup of authentication system based 
+     *                    on LDAP / AD.
+     *                    The user should be authorized to encrypt.
+     * (3) svcpassword -  SKCE service credentials : password of the service 
+     *                    username specified above
+     * (4) protocol    -  U2F protocol version to comply with.
      * 
      * @param payload - U2F de-registration parameters in Json form. Should
      * contain username and randomid.
+     * 
      * @return - A Json in String format. The Json will have 3 key-value pairs;
      * 1. 'Response' : String, with a simple message telling if the process was
      * successful or not. 2. 'Message' : Empty string since there is no
      * cryptographic work involved in de-registration 3. 'Error' : String, with
      * error message incase something went wrong. Will be empty if successful.
      */
-    @POST
-    @Path("/" + skfsConstants.FIDO_METHOD_DEREGISTER)
-    @Consumes({"application/x-www-form-urlencoded"})
-    @Produces({"application/json"})
-    public String deregister(@FormParam("svcinfo") String svcinfo,
-            @FormParam("payload") String payload) {
+    @WebMethod(operationName = "deregister")
+    public String deregister(
+            @WebParam(name = "svcinfo") SKCEServiceInfoType svcinfo,
+            @WebParam(name = "payload") String payload) 
+    {
         //  Local variables       
         //  Service credentials
         String did;
@@ -742,12 +700,13 @@ public class SKFEServlet {
         
         //  SKCE domain id validation
         try {
-            SKCEServiceInfoType si = basicInputChecks("deregister", svcinfo);
-            did = Integer.toString(si.getDid());
-            svcusername = si.getSvcusername();
-            svcpassword = si.getSvcpassword();
-            protocol = si.getProtocol();
-
+            basicInputChecks("deregister", svcinfo);
+            
+            did             = Integer.toString(svcinfo.getDid());
+            svcusername     = svcinfo.getSvcusername();
+            svcpassword     = svcinfo.getSvcpassword();
+            protocol        = svcinfo.getProtocol();
+            
             skfsCommon.inputValidateSKCEDid(did);
         } catch (Exception ex) {
             return skfsCommon.buildDeregisterResponse(null, "", ex.getLocalizedMessage());
@@ -755,12 +714,12 @@ public class SKFEServlet {
         
         //  2. Input checks
         if (svcusername == null || svcusername.isEmpty()) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcusername");
-            return skfsCommon.buildDeregisterResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcusername");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildDeregisterResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         if (svcpassword == null) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcpassword");
-            return skfsCommon.buildDeregisterResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcpassword");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildDeregisterResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         //authenticate
         boolean isAuthorized;
@@ -774,7 +733,7 @@ public class SKFEServlet {
             skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0033", "");
             return skfsCommon.buildDeregisterResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0033"));
         }
-        return u2fHelperBean.deregister(did, protocol, payload);
+        return u2fHelper.deregister(did, protocol, payload);
     }
 
     /*
@@ -796,15 +755,18 @@ public class SKFEServlet {
      * a unique registered key for that user. This random id can be obtained by
      * calling getkeysinfo method.
      *
-     * @param svcinfo - Object that carries SKCE service information.
-     * Information bundled is :
+     * @param svcinfo   - Object that carries SKCE service information. 
+     *                    Information bundled is :
      * 
-     * (1) did - Unique identifier for a SKCE encryption domain (2) svcusername
-     * - SKCE service credentials : username requesting the service. The service
-     * credentials are looked up in the 'service' setup of authentication system
-     * based on LDAP / AD. The user should be authorized to encrypt. (3)
-     * svcpassword - SKCE service credentials : password of the service username
-     * specified above (4) protocol - U2F protocol version to comply with.
+     * (1) did          - Unique identifier for a SKCE encryption domain
+     * (2) svcusername  - SKCE service credentials : username requesting the 
+     *                    service. The service credentials are looked up in 
+     *                    the 'service' setup of authentication system based 
+     *                    on LDAP / AD.
+     *                    The user should be authorized to encrypt.
+     * (3) svcpassword -  SKCE service credentials : password of the service 
+     *                    username specified above
+     * (4) protocol    -  U2F protocol version to comply with.
      * 
      * @param payload - U2F de-activation parameters in Json form. Should
      * contain username and randomid.
@@ -814,12 +776,11 @@ public class SKFEServlet {
      * cryptographic work involved in de-activation 3. 'Error' : String, with
      * error message incase something went wrong. Will be empty if successful.
      */
-    @POST
-    @Path("/" + skfsConstants.FIDO_METHOD_DEACTIVATE)
-    @Consumes({"application/x-www-form-urlencoded"})
-    @Produces({"application/json"})
-    public String deactivate(@FormParam("svcinfo") String svcinfo,
-            @FormParam("payload") String payload) {
+    @WebMethod(operationName = "deactivate")
+    public String deactivate(
+            @WebParam(name = "svcinfo") SKCEServiceInfoType svcinfo,
+            @WebParam(name = "payload") String payload) 
+    {
         //  Local variables       
         //  Service credentials
         String did;
@@ -829,12 +790,13 @@ public class SKFEServlet {
         
         //  SKCE domain id validation
         try {
-            SKCEServiceInfoType si = basicInputChecks("deactivate", svcinfo);
-            did = Integer.toString(si.getDid());
-            svcusername = si.getSvcusername();
-            svcpassword = si.getSvcpassword();
-            protocol = si.getProtocol();
-
+            basicInputChecks("deactivate", svcinfo);
+            
+            did             = Integer.toString(svcinfo.getDid());
+            svcusername     = svcinfo.getSvcusername();
+            svcpassword     = svcinfo.getSvcpassword();
+            protocol        = svcinfo.getProtocol();
+            
             skfsCommon.inputValidateSKCEDid(did);
         } catch (Exception ex) {
             return skfsCommon.buildDeactivateResponse(null, "", ex.getLocalizedMessage());
@@ -842,12 +804,12 @@ public class SKFEServlet {
         
         //  2. Input checks
         if (svcusername == null || svcusername.isEmpty()) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcusername");
-            return skfsCommon.buildDeactivateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcusername");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildDeactivateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         if (svcpassword == null) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcpassword");
-            return skfsCommon.buildDeactivateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcpassword");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildDeactivateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         //authenticate
         boolean isAuthorized;
@@ -861,7 +823,7 @@ public class SKFEServlet {
             skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0033", "");
             return skfsCommon.buildDeactivateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0033"));
         }
-        return u2fHelperBean.deactivate(did, protocol, payload);
+        return u2fHelper.deactivate(did, protocol, payload);
     }
 
     /*
@@ -883,15 +845,18 @@ public class SKFEServlet {
      * the random id to point to a unique registered key for that user. This
      * random id can be obtained by calling getkeysinfo method.
      *
-     * @param svcinfo - Object that carries SKCE service information.
-     * Information bundled is :
+     * @param svcinfo   - Object that carries SKCE service information. 
+     *                    Information bundled is :
      * 
-     * (1) did - Unique identifier for a SKCE encryption domain (2) svcusername
-     * - SKCE service credentials : username requesting the service. The service
-     * credentials are looked up in the 'service' setup of authentication system
-     * based on LDAP / AD. The user should be authorized to encrypt. (3)
-     * svcpassword - SKCE service credentials : password of the service username
-     * specified above (4) protocol - U2F protocol version to comply with.
+     * (1) did          - Unique identifier for a SKCE encryption domain
+     * (2) svcusername  - SKCE service credentials : username requesting the 
+     *                    service. The service credentials are looked up in 
+     *                    the 'service' setup of authentication system based 
+     *                    on LDAP / AD.
+     *                    The user should be authorized to encrypt.
+     * (3) svcpassword -  SKCE service credentials : password of the service 
+     *                    username specified above
+     * (4) protocol    -  U2F protocol version to comply with.
      * 
      * @param payload - U2F activation parameters in Json form. Should contain
      * username and randomid.
@@ -901,12 +866,11 @@ public class SKFEServlet {
      * cryptographic work involved in activation 3. 'Error' : String, with error
      * message incase something went wrong. Will be empty if successful.
      */
-    @POST
-    @Path("/" + skfsConstants.FIDO_METHOD_ACTIVATE)
-    @Consumes({"application/x-www-form-urlencoded"})
-    @Produces({"application/json"})
-    public String activate(@FormParam("svcinfo") String svcinfo,
-            @FormParam("payload") String payload) {
+    @WebMethod(operationName = "activate")
+    public String activate(
+            @WebParam(name = "svcinfo") SKCEServiceInfoType svcinfo,
+            @WebParam(name = "payload") String payload) 
+    {
         //  Local variables       
         //  Service credentials
         String did;
@@ -916,25 +880,26 @@ public class SKFEServlet {
         
         //  SKCE domain id validation
         try {
-            SKCEServiceInfoType si = basicInputChecks("activate", svcinfo);
-            did = Integer.toString(si.getDid());
-            svcusername = si.getSvcusername();
-            svcpassword = si.getSvcpassword();
-            protocol = si.getProtocol();
-
+            basicInputChecks("activate", svcinfo);
+            
+            did             = Integer.toString(svcinfo.getDid());
+            svcusername     = svcinfo.getSvcusername();
+            svcpassword     = svcinfo.getSvcpassword();
+            protocol        = svcinfo.getProtocol();
+            
             skfsCommon.inputValidateSKCEDid(did);
         } catch (Exception ex) {
-            return skfsCommon.buildDeactivateResponse(null, "", ex.getLocalizedMessage());
+            return skfsCommon.buildActivateResponse(null, "", ex.getLocalizedMessage());
         }
         
         //  2. Input checks
         if (svcusername == null || svcusername.isEmpty()) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcusername");
-            return skfsCommon.buildActivateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcusername");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildActivateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         if (svcpassword == null) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcpassword");
-            return skfsCommon.buildActivateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcpassword");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildActivateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         //authenticate
         boolean isAuthorized;
@@ -948,7 +913,7 @@ public class SKFEServlet {
             skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0033", "");
             return skfsCommon.buildActivateResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0033"));
         }
-        return u2fHelperBean.activate(did, protocol, payload);
+        return u2fHelper.activate(did, protocol, payload);
     }
 
     /*
@@ -973,17 +938,20 @@ public class SKFEServlet {
      * used (last modified) from, a random id (which has a time-to-live) that
      * has to be sent back as a token during de-registration.
      *
-     * @param svcinfo - Object that carries SKCE service information.
-     * Information bundled is :
+     * @param svcinfo   - Object that carries SKCE service information. 
+     *                    Information bundled is :
      * 
-     * (1) did - Unique identifier for a SKCE encryption domain (2) svcusername
-     * - SKCE service credentials : username requesting the service. The service
-     * credentials are looked up in the 'service' setup of authentication system
-     * based on LDAP / AD. The user should be authorized to encrypt. (3)
-     * svcpassword - SKCE service credentials : password of the service username
-     * specified above (4) protocol - U2F protocol version to comply with.
+     * (1) did          - Unique identifier for a SKCE encryption domain
+     * (2) svcusername  - SKCE service credentials : username requesting the 
+     *                    service. The service credentials are looked up in 
+     *                    the 'service' setup of authentication system based 
+     *                    on LDAP / AD.
+     *                    The user should be authorized to encrypt.
+     * (3) svcpassword -  SKCE service credentials : password of the service 
+     *                    username specified above
+     * (4) protocol    -  U2F protocol version to comply with.
      * 
-     * @param payload - Json input containing required data
+     * @param payload - username
      * @return - A Json in String format. The Json will have 3 key-value pairs;
      * 1. 'Response' : A Json array, each entry signifying metadata of a key
      * registered; Metadata includes randomid and its time-to-live, creation and
@@ -992,12 +960,11 @@ public class SKFEServlet {
      * String, with error message incase something went wrong. Will be empty if
      * successful.
      */
-    @POST
-    @Path("/" + skfsConstants.FIDO_METHOD_GETKEYSINFO)
-    @Consumes({"application/x-www-form-urlencoded"})
-    @Produces({"application/json"})
-    public String getkeysinfo(@FormParam("svcinfo") String svcinfo,
-            @FormParam("payload") String payload) {
+    @WebMethod(operationName = "getkeysinfo")
+    public String getkeysinfo(
+            @WebParam(name = "svcinfo") SKCEServiceInfoType svcinfo,
+            @WebParam(name = "payload") String payload) 
+    {
         //  Local variables       
         //  Service credentials
         String did;
@@ -1007,25 +974,26 @@ public class SKFEServlet {
         
         //  SKCE domain id validation
         try {
-            SKCEServiceInfoType si = basicInputChecks("getkeysinfo", svcinfo);
-            did = Integer.toString(si.getDid());
-            svcusername = si.getSvcusername();
-            svcpassword = si.getSvcpassword();
-            protocol = si.getProtocol();
-
+            basicInputChecks("getkeysinfo", svcinfo);
+            
+            did             = Integer.toString(svcinfo.getDid());
+            svcusername     = svcinfo.getSvcusername();
+            svcpassword     = svcinfo.getSvcpassword();
+            protocol        = svcinfo.getProtocol();
+            
             skfsCommon.inputValidateSKCEDid(did);
         } catch (Exception ex) {
-            return skfsCommon.buildDeactivateResponse(null, "", ex.getLocalizedMessage());
+            return skfsCommon.buildGetKeyInfoResponse(null, "", ex.getLocalizedMessage());
         }
         
         //  2. Input checks
         if (svcusername == null || svcusername.isEmpty()) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcusername");
-            return skfsCommon.buildGetKeyInfoResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcusername");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildGetKeyInfoResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         if (svcpassword == null) {
-            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " svcpassword");
-            return skfsCommon.buildGetKeyInfoResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " svcpassword");
+            skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0002", " credential");
+            return skfsCommon.buildGetKeyInfoResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0002") + " credential");
         }
         //authenticate
         boolean isAuthorized;
@@ -1039,7 +1007,7 @@ public class SKFEServlet {
             skfsLogger.log(skfsConstants.SKFE_LOGGER,Level.SEVERE, "FIDO-ERR-0033", "");
             return skfsCommon.buildGetKeyInfoResponse(null, "", skfsCommon.getMessageProperty("FIDO-ERR-0033"));
         }
-        return u2fHelperBean.getkeysinfo(did, protocol, payload);
+        return u2fHelper.getkeysinfo(did, protocol, payload);
     }
 
     /*
@@ -1063,25 +1031,24 @@ public class SKFEServlet {
      *
      * This method is *Not implemented yet*
      *
-     * @param svcinfo - Object that carries SKCE service information.
-     * Information bundled is :
+     * @param svcinfo   - Object that carries SKCE service information. 
+     *                    Information bundled is :
      * 
-     * (1) did - Unique identifier for a SKCE encryption domain (2) svcusername
-     * - SKCE service credentials : username requesting the service. The service
-     * credentials are looked up in the 'service' setup of authentication system
-     * based on LDAP / AD. The user should be authorized to encrypt. (3)
-     * svcpassword - SKCE service credentials : password of the service username
-     * specified above (4) protocol - U2F protocol version to comply with.
-     * @param payload
+     * (1) did          - Unique identifier for a SKCE encryption domain
+     * (2) svcusername  - SKCE service credentials : username requesting the 
+     *                    service. The service credentials are looked up in 
+     *                    the 'service' setup of authentication system based 
+     *                    on LDAP / AD.
+     *                    The user should be authorized to encrypt.
+     * (3) svcpassword -  SKCE service credentials : password of the service 
+     *                    username specified above
+     * (4) protocol    -  U2F protocol version to comply with.
      * 
      * @return
      */
-    @POST
-    @Path("/" + skfsConstants.FIDO_METHOD_GETSERVERINFO)
-    @Consumes({"application/x-www-form-urlencoded"})
-    @Produces({"application/json"})
-    public String getserverinfo(@FormParam("svcinfo") String svcinfo,
-            @FormParam("payload") String payload) {
+    @WebMethod(operationName = "getserverinfo")
+    public String getserverinfo(
+            @WebParam(name = "svcinfo") SKCEServiceInfoType svcinfo) {
         return "not implemented yet";
     }
 }
